Move saveLearnerPluginState into IPluginRuntimeContext
[#165291182]
