+++ lara-plugin-api/src/helpers/plugin-runtime-context.ts
@@ -2,6 +2,7 @@ import {
   IPluginRuntimeContext, IJwtResponse, IClassInfo
 } from "../api/types";
 import { generateEmbeddableRuntimeContext } from "./embeddable-runtime-context";
+import * as $ from "jquery";
 
 /** Data generated by LARA, passed to .initPlugin function, and then used to generate final IPluginRuntimeContext */
 export interface IPluginContext {
@@ -62,6 +63,22 @@ export interface IEmbeddableContext {
   clickToPlayId: string | null;
 }
 
+export const saveLearnerPluginState = (learnerStateSaveUrl: string, state: string): Promise<string> => {
+  return new Promise((resolve, reject) => {
+    $.ajax({
+      url: learnerStateSaveUrl,
+      type: "PUT",
+      data: { state },
+      success(data) {
+        resolve(data);
+      },
+      error(jqXHR, errText, err) {
+        reject(err);
+      }
+    });
+  });
+};
+
 const getFirebaseJwt = (firebaseJwtUrl: string, appName: string): Promise<IJwtResponse> => {
   const appSpecificUrl = firebaseJwtUrl.replace("_FIREBASE_APP_", appName);
   return fetch(appSpecificUrl, {method: "POST"})
@@ -98,6 +115,7 @@ export const generatePluginRuntimeContext = (context: IPluginContext): IPluginRu
     runId: context.runId,
     remoteEndpoint: context.remoteEndpoint,
     userEmail: context.userEmail,
+    saveLearnerPluginState: (state: string) => saveLearnerPluginState(context.learnerStateSaveUrl, state),
     getClassInfo: () => getClassInfo(context.classInfoUrl),
     getFirebaseJwt: (appName: string) => getFirebaseJwt(context.firebaseJwtUrl, appName),
     wrappedEmbeddable: context.wrappedEmbeddable ? generateEmbeddableRuntimeContext(context.wrappedEmbeddable) : null
